Add generic api code

+++ MVVMProducts/Screens/Product/Model/Product.swift
@@ -10,14 +10,14 @@ import Foundation
 struct Product: Decodable {
     let id: Int
     let title: String
-    let price:  Float
+    let price:  Double
     let description: String
     let category: String
     let image: String
     let rating: Rate
 }
 struct Rate: Decodable {
-    let rate: Float
+    let rate: Double
     let count: Int
 }
